backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-core>=0.2.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langgraph>=0.2.0
openai>=1.40.0
anthropic>=0.34.0
tiktoken>=0.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
tenacity>=8.2.3
python-dotenv>=1.0.1
pytest>=7.4.0
pytest-asyncio>=0.21.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
